requirements.txt
pandas>=0.12.0
matplotlib>=1.3.1
numpy>=1.7.1
Pytorch >= 0.3.1